feat: small improvements for ux
